refactor axios methods
