+++ utilities/helpers/gameProcess.js
