+++ utilities/helpers/startGame.js
